added the sql group

+++ movies-restful-service/src/main/java/com/movies/repositry/MoviesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MoviesRepository extends CrudRepository<Movie, Integer> {
+public interface MoviesRepository extends CrudRepository<Movie, Long> {
 
     @Query("select m from Movie m where m.movie_name like %?1%")
     List<Movie> findByMovieName(String movieName);
